Update script to ping wallpaper server, then wait an additional 40 seconds before downloading
I've been having issues where, when the script tries to download the XML metadata, the server is down, but comes up within 20-30 seconds. My theory is the server is set to shut down when not in use, and needs time to come back online when a request is made. My very hacky solution is to just ping the server, then wait 40 seconds, *then* request the XML. Works fine.