Remove server ping, as it seems to work normally again
This script contained code to ping the server containing the Bing wallpaper metadata, and wait for it to wake up before obtaining the wallpaper. This is because of behaviour I noticed in January where the server was not always active. Now, that seems to have been fixed, so I have removed the delay.